+++ packages/api/src/platforms/vtex/clients/commerce/types/VtexId.ts
@@ -1,0 +1,11 @@
+export interface VtexIdResponse {
+  authStatus: string
+  id: string
+  user: string
+  account: string
+  audience: string
+  tokenType: string
+  customerId: string
+  unitName: string
+  isRepresentative: boolean
+}