+++ packages/api/src/platforms/vtex/resolvers/facet.ts
@@ -11,8 +11,12 @@ export type Root = Facet
 
 export const StoreFacet: Record<string, Resolver<Root>> = {
   __resolveType: ({ type, values }) => {
-    if (type !== 'TEXT') {
+    if (type !== 'TEXT' && type !== 'DELIVERY') {
       return 'StoreFacetRange'
+    }
+
+    if (!values || values.length === 0) {
+      return 'StoreFacetBoolean'
     }
 
     if (values.every((value) => (value as FacetValueRange).range)) {
@@ -51,7 +55,7 @@ export const StoreFacetRange: Record<
     )
 
     const facet = min(values, (a, b) => a.range.from - b.range.from)
-    const globalMin = facet?.range.from ?? 0
+    const globalMin = facet?.range?.from ?? 0
 
     return {
       selected: selectedRange?.[0] ?? globalMin,
@@ -71,7 +75,7 @@ export const StoreFacetRange: Record<
     )
 
     const facet = min(values, (a, b) => b.range.to - a.range.to)
-    const globalMax = facet?.range.to ?? 0
+    const globalMax = facet?.range?.to ?? 0
 
     return {
       selected: selectedRange?.[1] ?? globalMax,