+++ packages/sdk/src/session/index.ts
@@ -39,6 +39,13 @@ export interface Person {
 
 export interface B2bInfo {
   customerId: string
+  isRepresentative: boolean
+  unitName: string
+  unitId: string
+  firstName: string
+  lastName: string
+  userName: string
+  userEmail: string
 }
 
 export interface Session {